pyproject.toml
[tool.poetry]
name = "mcp_skeleton"
version = "0.1.0"
description = "Skeleton framework for building MCP (Model Context Protocol) servers"
authors = ["Your Name <your.email@example.com>"]
packages = [
    {include = "app"},
]

[tool.poetry.dependencies]
python = ">=3.11,<4"
requests = "^2.31.0"
mcp = "^1.3.0"
uvicorn = "^0.29.0"
starlette = "^0.36.3"
python-dotenv = "^1.0.0"
pydantic = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mcp-server = "app.__main__:main" 